organized file structure, updated astro dependency
