+++ package.json
@@ -13,7 +13,7 @@
         "@astrojs/mdx": "^4.3.0",
         "@astrojs/svelte": "^7.1.0",
         "@tailwindcss/vite": "^4.1.8",
-        "astro": "^5.9.1",
+        "astro": "^5.10.0",
         "caniuse-lite": "^1.0.30001698",
         "moment": "^2.30.1",
         "moment-timezone": "^0.6.0",
