+++ src/components/experience/MoreExperience.astro
@@ -1,6 +1,6 @@
 ---
-import Card from "../elements/Card.astro";
-import Line from "../elements/NetworkLine.astro";
+import Card from "../global/Card.astro";
+import Line from "../global/NetworkLine.astro";
 ---
 
 <div id="Experience" class="h-fit">
