+++ src/components/global/FooterComponent.astro
