+++ src/components/global/NavBar.astro
