+++ src/components/global/NetworkLine.astro
