+++ src/components/projects/ProjectsComponent.astro
@@ -1,9 +1,9 @@
 ---
 import TaskWidget from "./TaskWidget.svelte";
 import { Image } from "astro:assets";
-import CoworkingGuide from "./images/projects/coworking-guide.webp";
-import Wnrs from "./images/projects/wnrs.webp";
-import TwitchPlays from "./images/projects/twitchplays.webp";
+import CoworkingGuide from "../images/projects/coworking-guide.webp";
+import Wnrs from "../images/projects/wnrs.webp";
+import TwitchPlays from "../images/projects/twitchplays.webp";
 ---
 
 <div
