+++ src/components/sections/skills/MoreSkills.astro
@@ -1,0 +1,103 @@
+---
+import Card from "../global/Card.astro";
+import Line from "../global/NetworkLine.astro";
+---
+
+<div id="Experience" class="h-fit">
+	<div class="flex flex-col items-center">
+		<div class="w-[90%] md:w-[600px]">
+			<Line height="50px" />
+			<div class="mx-auto w-[90%] md:w-[600px]">
+				<Card primaryColor="#9D76C1">
+					<div class="px-8 py-8">
+						<div class="justify-left flex w-full flex-col">
+							<h2 class="text-2xl font-bold">Content Creation</h2>
+						</div>
+						<hr class="my-3 w-full" />
+						<p class="text-left">
+							Since I was little, I loved watching YouTube videos
+							and the idea of bringing joy and laughter to a wider
+							audience. Thus, I began my journey in content
+							creation through YouTube videos and Twitch
+							streaming. I typically do coworking & study streams
+							to keep myself accountable and encourage
+							productivity amongst viewers. I also do software and
+							game development streams where I showcase my
+							workflow.
+						</p>
+						<div
+							class="flex w-full flex-row flex-wrap justify-start"
+						>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-[#141414] px-3 py-1 text-white transition-colors duration-300 hover:bg-[#141414] hover:text-white md:bg-gray-500"
+							>
+								OBS
+							</div>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-gray-50 px-3 py-1 text-black transition-colors duration-300 hover:bg-gray-50 hover:text-black md:bg-gray-500 md:text-white"
+							>
+								DaVinci Resolve
+							</div>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-purple-500 px-3 py-1 text-white transition-colors duration-300 hover:bg-purple-500 md:bg-gray-500"
+							>
+								Twitch
+							</div>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-[#ff0000] px-3 py-1 text-white transition-colors duration-300 hover:bg-[#ff0000] hover:text-white md:bg-gray-500"
+							>
+								YouTube
+							</div>
+						</div>
+					</div>
+				</Card>
+			</div>
+			<Line height="50px" />
+			<div class="mx-auto w-[90%] md:w-[600px]">
+				<Card primaryColor="#9D76C1">
+					<div class="px-8 py-8">
+						<div class="justify-left flex w-full flex-col">
+							<h2 class="text-2xl font-bold">
+								Mobile Development
+							</h2>
+						</div>
+						<hr class="my-3 w-full" />
+						<p class="text-left">
+							In my Mobile App Development unit in University, I
+							am tasked to develop Android apps using Android
+							Studio. The final app project I had developed was a
+							LLM-powered quiz app with features such as choosing
+							your own topics, viewing your question attempt
+							history afterwards, and upgrading your tier with
+							Google Pay.
+						</p>
+						<p class="text-left">
+							The marking tutor was impressed with my work, and
+							offered me to join a research lab because I had good
+							potential.
+						</p>
+						<div
+							class="mt-5 flex w-full flex-row flex-wrap justify-start"
+						>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-[#106539] px-3 py-1 text-white transition-colors duration-300 hover:bg-[#106539] hover:text-white md:bg-gray-500"
+							>
+								Android
+							</div>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-[#ED8B00] px-3 py-1 text-white transition-colors duration-300 hover:bg-[#ED8B00] md:bg-gray-500"
+							>
+								Java
+							</div>
+							<div
+								class="my-1 mr-1 w-fit cursor-default rounded-full bg-[#7F52FF] px-3 py-1 text-white transition-colors duration-300 hover:bg-[#7F52FF] md:bg-gray-500"
+							>
+								Kotlin
+							</div>
+						</div>
+					</div>
+				</Card>
+			</div>
+		</div>
+	</div>
+</div>
