+++ src/components/skills/SkillsComponent.astro
@@ -1,6 +1,6 @@
 ---
-import Card from "../elements/Card.astro";
-import Line from "../elements/NetworkLine.astro";
+import Card from "../global/Card.astro";
+import Line from "../global/NetworkLine.astro";
 ---
 
 <div class="h-fit w-full pt-20">
