+++ src/components/what-i-do/WhatIDoComponent.astro
@@ -1,5 +1,5 @@
 ---
-import Card from "../elements/Card.astro";
+import Card from "../global/Card.astro";
 ---
 
 <div class="flex w-full flex-col items-center justify-center">
