+++ src/pages/experience.astro
@@ -1,19 +1,19 @@
 ---
-import NavBar from "../components/sections/NavBar.astro";
+import NavBar from "../components/global/NavBar.astro";
 import Layout from "../layouts/Layout.astro";
-import EducationComponent from "../components/sections/ExperienceComponent.astro";
-import MoreExperience from "../components/sections/MoreExperience.astro";
-import Footer from "../components/sections/FooterComponent.astro";
+import EducationComponent from "../components/experience/ExperienceComponent.astro";
+import MoreExperience from "../components/experience/MoreExperience.astro";
+import Footer from "../components/global/FooterComponent.astro";
 ---
 
 <Layout
-    title="Experience"
-    description="Find out RythonDev's previous education"
+	title="Experience"
+	description="Find out RythonDev's previous education"
 >
-    <NavBar />
-    <div id="Experience" class="py-20">
-        <EducationComponent />
-        <MoreExperience />
-    </div>
-    <Footer />
+	<NavBar />
+	<div id="Experience" class="py-20">
+		<EducationComponent />
+		<MoreExperience />
+	</div>
+	<Footer />
 </Layout>
