+++ src/pages/index.astro
@@ -1,61 +1,61 @@
 ---
-import NavBar from "../components/sections/NavBar.astro";
+import NavBar from "../components/global/NavBar.astro";
 import Layout from "../layouts/Layout.astro";
-import NetworkLine from "../components/elements/NetworkLine.astro";
-import WhatIDo from "../components/sections/WhatIDoComponent.astro";
-import Intro from "../components/sections/IntroComponent.svelte";
-import Projects from "../components/sections/ProjectsComponent.astro";
-import Experience from "../components/sections/ExperienceComponent.astro";
-import Skills from "../components/sections/SkillsComponent.astro";
-import Contact from "../components/sections/ContactComponent.svelte";
-import Card from "../components/elements/Card.astro";
-import Footer from "../components/sections/FooterComponent.astro";
+import NetworkLine from "../components/global/NetworkLine.astro";
+import WhatIDo from "../components/what-i-do/WhatIDoComponent.astro";
+import Intro from "../components/home/IntroComponent.svelte";
+import Projects from "../components/projects/ProjectsComponent.astro";
+import Experience from "../components/experience/ExperienceComponent.astro";
+import Skills from "../components/skills/SkillsComponent.astro";
+import Contact from "../components/contact/ContactComponent.svelte";
+import Card from "../components/global/Card.astro";
+import Footer from "../components/global/FooterComponent.astro";
 ---
 
 <Layout
-    title="Ryan's Portfolio"
-    description="Ryan is a software developer, streamer and Computer Science undergraduate. Learn more about him and his skills here!"
+	title="Ryan's Portfolio"
+	description="Ryan is a software developer, streamer and Computer Science undergraduate. Learn more about him and his skills here!"
 >
-    <NavBar />
-    <div id="home" class="flex items-center justify-center pt-20">
-        <div class="w-[90%] max-w-[1080px] lg:w-1/2 xl:w-1/2">
-            <Card
-                bgColor="#252525"
-                primaryColor="#BEADFA"
-                className="animate-fadeIn"
-            >
-                <Intro client:load />
-            </Card>
-        </div>
-    </div>
-    <NetworkLine height="70px" />
-    <div id="what-i-do">
-        <div class="mx-auto mb-10 h-[1px] w-1/3 bg-[#404040]"></div>
-        <WhatIDo />
-    </div>
-    <div id="projects">
-        <div class="pt-20">
-            <h1 class="mb-10 text-center text-5xl">Projects</h1>
-            <Projects />
-        </div>
-    </div>
-    <div id="skills">
-        <Skills />
-    </div>
-    <div id="education" class="py-20">
-        <Experience />
-    </div>
-    <div id="contact">
-        <Contact />
-    </div>
-    <Footer />
+	<NavBar />
+	<div id="home" class="flex items-center justify-center pt-20">
+		<div class="w-[90%] max-w-[1080px] lg:w-1/2 xl:w-1/2">
+			<Card
+				bgColor="#252525"
+				primaryColor="#BEADFA"
+				className="animate-fadeIn"
+			>
+				<Intro client:load />
+			</Card>
+		</div>
+	</div>
+	<NetworkLine height="70px" />
+	<div id="what-i-do">
+		<div class="mx-auto mb-10 h-[1px] w-1/3 bg-[#404040]"></div>
+		<WhatIDo />
+	</div>
+	<div id="projects">
+		<div class="pt-20">
+			<h1 class="mb-10 text-center text-5xl">Projects</h1>
+			<Projects />
+		</div>
+	</div>
+	<div id="skills">
+		<Skills />
+	</div>
+	<div id="education" class="py-20">
+		<Experience />
+	</div>
+	<div id="contact">
+		<Contact />
+	</div>
+	<Footer />
 </Layout>
 <script is:inline>
-    location.search.includes("?uwu") &&
-        document.documentElement.classList.add("uwu");
+	location.search.includes("?uwu") &&
+		document.documentElement.classList.add("uwu");
 </script>
 <style is:global>
-    .uwu #uwu {
-        display: flex;
-    }
+	.uwu #uwu {
+		display: flex;
+	}
 </style>
