+++ src/pages/skills.astro
@@ -1,9 +1,9 @@
 ---
-import NavBar from "../components/sections/NavBar.astro";
+import NavBar from "../components/global/NavBar.astro";
 import Layout from "../layouts/Layout.astro";
-import SkillsComponent from "../components/sections/SkillsComponent.astro";
-import MoreSkills from "../components/sections/MoreSkills.astro";
-import Footer from "../components/sections/FooterComponent.astro";
+import SkillsComponent from "../components/skills/SkillsComponent.astro";
+import MoreSkills from "../components/skills/MoreSkills.astro";
+import Footer from "../components/global/FooterComponent.astro";
 ---
 
 <Layout
