+++ src/pages/socials.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from "../layouts/Layout.astro";
-import StreamScheduleComponent from "../components/sections/StreamSchedule.svelte";
-import Footer from "../components/sections/FooterComponent.astro";
+import StreamScheduleComponent from "../components/socials/StreamSchedule.svelte";
+import Footer from "../components/global/FooterComponent.astro";
 import { Image } from "astro:assets";
 import profile from "./_images/profile.webp";
 import WebsiteIcon from "./_images/website-icon.svg";
