+++ src/pages/what-i-do.astro
@@ -1,10 +1,10 @@
 ---
-import NavBar from "../components/sections/NavBar.astro";
+import NavBar from "../components/global/NavBar.astro";
 import Layout from "../layouts/Layout.astro";
-import WhatIDoComponent from "../components/sections/WhatIDoComponent.astro";
-import Card from "../components/elements/Card.astro";
-import Line from "../components/elements/NetworkLine.astro";
-import Footer from "../components/sections/FooterComponent.astro";
+import WhatIDoComponent from "../components/what-i-do/WhatIDoComponent.astro";
+import Card from "../components/global/Card.astro";
+import Line from "../components/global/NetworkLine.astro";
+import Footer from "../components/global/FooterComponent.astro";
 ---
 
 <Layout
